+++ src/Damage/ProbabilisticDamageModel.cc
@@ -23,6 +23,7 @@
 #include "DataBase/State.hh"
 #include "DataBase/StateDerivatives.hh"
 #include "DataBase/ReplaceState.hh"
+#include "DataBase/updateStateFields.hh"
 #include "Hydro/HydroFieldNames.hh"
 #include "Field/FieldList.hh"
 #include "Boundary/Boundary.hh"
@@ -100,7 +101,7 @@ ProbabilisticDamageModel<Dimension>::
 }
 
 //------------------------------------------------------------------------------
-// initializeProblemStartup
+// initializeProblemStartupDependencies
 //
 // After all initial state has been initialize (node positions, masses, etc),
 // but before we try to run any physics cycles.  This is when we initialize a
@@ -109,7 +110,9 @@ ProbabilisticDamageModel<Dimension>::
 template<typename Dimension>
 void
 ProbabilisticDamageModel<Dimension>::
-initializeProblemStartup(DataBase<Dimension>& dataBase) {
+initializeProblemStartupDependencies(DataBase<Dimension>& dataBase,
+                                     State<Dimension>& state,
+                                     StateDerivatives<Dimension>& derivs) {
 
   // How many points are actually being damaged?
   // We have to be careful to use an unsigned size_t here due to overflow
@@ -135,8 +138,11 @@ initializeProblemStartup(DataBase<Dimension>& dataBase) {
   // Compute the initial volumes and random seeds for each node.  We hash the
   // morton index of each point with the global seed value to create unique
   // but reproducible seeds for each points random number generator.
-  const auto& mass = nodes.mass();
-  const auto& rho = nodes.massDensity();
+  auto buildKey = [&](const std::string& fkey) -> std::string { return StateBase<Dimension>::buildFieldKey(fkey, nodes.name()); };
+  const auto& mass = state.field(buildKey(HydroFieldNames::mass), 0.0);
+  const auto& rho = (state.registered(buildKey(SolidFieldNames::porositySolidDensity)) ?
+                     state.field(buildKey(SolidFieldNames::porositySolidDensity), 0.0) :
+                     state.field(buildKey(HydroFieldNames::massDensity), 0.0));
   const auto  nlocal = nodes.numInternalNodes();
   vector<uniform_random> randomGenerators(nlocal);
 #pragma omp parallel for
@@ -221,21 +227,13 @@ initializeProblemStartup(DataBase<Dimension>& dataBase) {
     }
   }
 
-  // We need a bunch of state to set the Youngs modulus and longitudinal sound speed
-  const auto& eps = nodes.specificThermalEnergy();
-  const auto& D = nodes.damage();
-  Field<Dimension, Scalar> P("P", nodes), K("K", nodes), mu("mu", nodes);
-  nodes.equationOfState().setPressure(P, rho, eps);
-  if (nodes.strengthModel().providesBulkModulus()) {
-    nodes.strengthModel().bulkModulus(K, rho, eps);
-  } else {
-    nodes.equationOfState().setBulkModulus(K, rho, eps);
-  }
-  nodes.strengthModel().shearModulus(mu, rho, eps, P, D);
-
-  // Set the initial values for Youngs modulus and the longitudinal sound speed
-  nodes.YoungsModulus(mYoungsModulus, K, mu);
-  nodes.longitudinalSoundSpeed(mLongitudinalSoundSpeed, rho, K, mu);
+  // Set the moduli.
+  updateStateFields(HydroFieldNames::pressure, state, derivs);
+  updateStateFields(SolidFieldNames::bulkModulus, state, derivs);
+  updateStateFields(SolidFieldNames::shearModulus, state, derivs);
+  updateStateFields(SolidFieldNames::yieldStrength, state, derivs);
+  updateStateFields(SolidFieldNames::YoungsModulus, state, derivs);
+  updateStateFields(SolidFieldNames::longitudinalSoundSpeed, state, derivs);
 }
 
 //------------------------------------------------------------------------------