+++ src/Field/FieldList.hh
@@ -247,6 +247,7 @@ public:
   // The number of fields in the FieldList.
   unsigned numFields() const;
   unsigned size() const;
+  bool empty() const;
 
   // The number of nodes in the FieldList.
   unsigned numNodes() const;